+++ browser-service/src/main/java/com/platon/browser/res/address/DetailsRPPlanResp.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.platon.browser.config.CustomLatSerializer;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 /**
  *  地址详情锁仓子结构体返回对象
@@ -13,14 +14,14 @@ import java.math.BigDecimal;
  *  @data 2019年8月31日
  */
 public class DetailsRPPlanResp {
-	private Long epoch;         //锁仓周期
+	private BigInteger epoch;         //锁仓周期
     private BigDecimal amount;      //锁定金额
-    private Long blockNumber;   //锁仓周期对应快高  结束周期 * epoch  
+    private BigInteger blockNumber;   //锁仓周期对应快高  结束周期 * epoch  
     private Long estimateTime;   //预计时间
-	public Long getEpoch() {
+	public BigInteger getEpoch() {
 		return epoch;
 	}
-	public void setEpoch(Long epoch) {
+	public void setEpoch(BigInteger epoch) {
 		this.epoch = epoch;
 	}
 	@JsonSerialize(using = CustomLatSerializer.class)
@@ -30,10 +31,10 @@ public class DetailsRPPlanResp {
 	public void setAmount(BigDecimal amount) {
 		this.amount = amount;
 	}
-	public Long getBlockNumber() {
+	public BigInteger getBlockNumber() {
 		return blockNumber;
 	}
-	public void setBlockNumber(Long blockNumber) {
+	public void setBlockNumber(BigInteger blockNumber) {
 		this.blockNumber = blockNumber;
 	}
 	public Long getEstimateTime() {
